requirements.txt
asgiref==3.9.1
black==25.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.10.1
Django==5.2.4
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
flake8==7.3.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
